# packages/qn-addon/openapi.yaml
openapi: 3.0.3
info:
  title: Veil Privacy Suite — QuickNode Add-On
  description: |
    End-to-end encryption, secret sharing, and ZK compression for Solana dApps.
    MEV-protect orders, manage encrypted state, and gate features by on-chain reputation.

    ## Features
    - **NaCl Box Encryption** — Curve25519-XSalsa20-Poly1305 for end-to-end encrypted messaging
    - **Shamir Secret Sharing** — M-of-N threshold schemes for multi-party key management
    - **Order Encryption** — MEV-protected swap orders for Solana DEX integrations
    - **ZK Compression** — Light Protocol integration for 99% state cost reduction
    - **Reputation Tiers** — FairScore-based feature gating with 5 tier levels
    - **Payload Serialization** — Type-safe binary encoding for on-chain data

    ## Authentication
    API endpoints use rate limiting (100 requests/minute).
    Provisioning endpoints require HTTP Basic Auth.
  version: 0.1.0
  contact:
    name: Veil Privacy Suite Support
    email: support@veil-privacy.dev
  license:
    name: MIT
servers:
  - url: http://localhost:3030
    description: Local development
  - url: https://your-addon.example.com
    description: Production
...
paths:
...
  /v1/threshold/split:
    post:
      tags: [Threshold]
      summary: Split secret into shares
      description: |
        Splits a 32-byte secret into N shares using Shamir's Secret Sharing.
        Any M (threshold) shares can reconstruct the original secret.
        Uses polynomial evaluation over a 256-bit prime field.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [secret, threshold, totalShares]
              properties:
                secret:
                  type: string
                  format: base64
                  description: 32-byte secret to split (base64)
                threshold:
                  type: integer
                  minimum: 2
                  description: Minimum shares needed to reconstruct (M)
                totalShares:
                  type: integer
                  minimum: 2
                  maximum: 255
                  description: Total shares to generate (N)
      responses:
        '200':
          description: Generated shares
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  shares:
                    type: array
                    items:
                      $ref: '#/components/schemas/SecretShare'
        '400':
          $ref: '#/components/responses/BadRequest'
  /v1/threshold/combine:
    post:
      tags: [Threshold]
      summary: Combine shares to reconstruct secret
      description: Reconstructs the original secret from threshold shares using Lagrange interpolation.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [shares]
              properties:
                shares:
                  type: array
                  minItems: 2
                  items:
                    $ref: '#/components/schemas/SecretShare'
      responses:
        '200':
          description: Reconstructed secret
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  secret:
                    $ref: '#/components/schemas/EncodedBytes'
        '400':
          $ref: '#/components/responses/BadRequest'
  /v1/threshold/verify:
    post:
      tags: [Threshold]
      summary: Verify share consistency
      description: |
        Verifies that a set of shares are consistent by attempting partial
        reconstruction with different subsets. If two different subsets of
        threshold shares produce the same secret, the shares are valid.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [shares, threshold]
              properties:
                shares:
                  type: array
                  minItems: 2
                  items:
                    $ref: '#/components/schemas/SecretShare'
                threshold:
                  type: integer
                  minimum: 2
                  description: Expected threshold for these shares
      responses:
        '200':
          description: Verification result
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  valid:
                    type: boolean
                    description: Whether shares are consistent
                  sharesProvided:
                    type: integer
                  threshold:
                    type: integer
        '400':
          $ref: '#/components/responses/BadRequest'

  # === ORDERS ENDPOINTS ===
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: HTTP Basic Auth for provisioning endpoints
...
  schemas:
    EncodedBytes:
      type: object
      properties:
        base64:
          type: string
          format: base64
        hex:
          type: string
...
    SecretShare:
      type: object
      properties:
        index:
          type: integer
          description: Share index (1-based)
        value:
          type: string
          format: base64
          description: Share value (base64)
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'